+++ cljs/src/hive_jam/core.cljs
@@ -1008,10 +1008,12 @@
       (dom/div #js {:style #js {:fontFamily "monospace"
                                 :background (:background theme)}}
                (om/build error-view (:errors cursor))
-               (om/build hive-view {:cursor (:hive cursor)
-                                    :save-state-ch (:save-state-ch state)
-                                    :load-state-ch (:load-state-ch state)})
-               (om/build audio-view (:audio cursor))
+               (when-not (empty? (config "StateFile"))
+                 (om/build hive-view {:cursor (:hive cursor)
+                                      :save-state-ch (:save-state-ch state)
+                                      :load-state-ch (:load-state-ch state)}))
+               (when (config "EnableUiAudio")
+                 (om/build audio-view (:audio cursor)))
                (om/build grid-view {:id "root" :beat-cursors (:beat-cursors cursor)} {:state state})
                (om/build console-view (:console cursor))))))