+++ RELEASE_NOTES.md
@@ -6,7 +6,7 @@
 - Grid deep copy in track type UI dropdown
 - Increase/decrease grid resolution
 
-Full list of issues: [Milestone v0.2](https://github.com/josephburnett/sonic-jam/issues?q=is%3Aissue+milestone%3Av0.2+is%3Aclosed)
+Full list of issues: [Milestone v0.2](https://github.com/josephburnett/hive-jam/issues?q=is%3Aissue+milestone%3Av0.2+is%3Aclosed)
 
 Tested with Sonic Pi v2.10